Implement RFile.mmapRead ##util
